CloudFiles can now request the list of containers for a user in a specific region by specifying the region on the API call.
Need to support a region specified (immutably) in the configuration. But
then allow the user to override it on a per method basis.

A client now has multiple Transports: one per region. We're not going
to have any of that "one client per region" nonsense.

But this introduces Waaaaay too much code driven by a single test...