When a method is added to a DocsLinkable class, and is preceded by a call to the #docs method, wrap this method, via an alias method chain, with a rescue clause that augments any Exception's message with the URL of the API docs for the API call in question and re-raise the Exception.
This is perhaps a bit too clever but, assuming I didn't screw up, this
neatly allows us to not only semantically document the individual API
calls with URL's to the API docs but also capture that same information
in the Exceptions.

Of course, I realize now that I'm missing that same information in the
comments in the SDK itself now. :(