Region now cascades down to children. * File, when making an API call, looks for its own region, then its container's, then its client's. * Container likewise looks for its own region then its client's
I kind of hate this. Though `region` may be the only piece of data that
requires these sorts of contortions.  For any other service that has
multiple layers of child objects, we're going to encounter this.

Really, it's a recursive tree climb up with the root node as a base
case. Maybe this can be simplified that way.  Or perhaps it should
simply be removed in favor of letting the user create a new service,
e.g., "CloudFiles.new", with each instance having its own region much
like how we currently handle this in Fog.