Don't try this at home, kids!
YARD registers developer-defined handlers to run *before* its own
handlers. This seems to cause my changes to the MethodObjects to become
overwritten--so I cheated.  I reversed the order of the registered
handlers.

Otherwise, all this does is add a Tag object to the Docstring object
attached to the MethodObject.  These objects are created during YARD's
process of traversing the AST resulting form the parse of the
code/comments.  This tag is treated like a "@see foo" line in the YARD
doc for the purposes of rendering the documentation.